requirements.txt
flask
torch
torchvision
Pillow
